CSSTUDIO-1936: relocated/renamed HtmlPreview to CommonmarkPreview
