+++ src/components/shared/CommonmarkPreview/index.js
@@ -1,0 +1,4 @@
+import { CommonmarkPreview } from "./CommonmarkPreview"
+
+export { CommonmarkPreview };
+export default CommonmarkPreview;
